requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
mysqlclient
python-dotenv
pillow
drf-yasg